update: make comment group independent
