+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 -   移除：播放页 接下来播放免过滤、播放结束免过滤
 -   新增：播放页 对相关视频暂存数据进行整体过滤
 -   更新：直播页 弹幕列表过滤功能，适配页面变化
+-   更新：动态页 净化功能适配新版评论区
+-   优化：评论区净化功能从各页面分离，成为全站通用板块
 -   优化：首页净化功能
 
 ## 3.10.4
