+++ package.json
@@ -17,11 +17,11 @@
         "@typescript-eslint/parser": "^6.21.0",
         "eslint": "^8.57.0",
         "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-prettier": "^5.1.3",
+        "eslint-plugin-prettier": "^5.2.1",
         "husky": "^8.0.3",
-        "prettier": "^3.2.5",
-        "sass": "^1.77.4",
-        "typescript": "^5.4.4",
+        "prettier": "^3.3.3",
+        "sass": "^1.77.8",
+        "typescript": "^5.5.4",
         "vite": "^4.5.3",
         "vite-plugin-monkey": "^3.5.2"
     }
