+++ src/rules/video.ts
@@ -5,7 +5,6 @@ import { matchAvidBvid, matchBvid, waitForEle } from '../utils/tool'
 import { isPageFestival, isPagePlaylist, isPageVideo } from '../utils/pageType'
 import { GM_getValue, GM_setValue, unsafeWindow } from '$'
 import URLCleanerInstance from '../utils/urlCleaner'
-import { Shadow } from '../utils/shadow'
 
 /** 宽屏模式监听 */
 let _isWide = unsafeWindow.isWide
@@ -36,27 +35,6 @@ if (isPageVideo() || isPagePlaylist()) {
         }
     })
 }
-
-// shadow DOM 评论区
-const shadow = new Shadow([
-    'bili-comments-header-renderer', // 评论区header(notice,编辑器)
-    'bili-comment-textarea', // 主编辑器、底部编辑器输入框
-    'bili-comment-renderer', // 一级评论
-    'bili-comment-user-info', // 一级二级评论用户信息
-    'bili-rich-text', // 一级二级评论内容
-    'bili-avatar', // 头像
-    'bili-photoswipe', // 全屏图片查看(笔记图片)
-    'bili-user-profile', // 用户卡片
-    'bili-comment-user-medal', // 昵称后勋章
-    'bili-comment-action-buttons-renderer', // 赞踩回复按钮区域
-    'bili-comment-reply-renderer', // 单个二级评论
-    // 'bili-comments', // 评论区板块
-    // 'bili-comments-notice', // 活动通知notice
-    // 'bili-comment-box', // 主编辑器、底部编辑器
-    // 'bili-comment-thread-renderer', // 评论（含回复）
-    // 'bili-comment-replies-renderer', // 多个二级评论
-    // 'bili-comment-user-sailing-card', // 一级评论右侧饰品
-])
 
 const disableAdjustVolume = () => {}
 
@@ -1561,656 +1539,6 @@ if (isPageVideo() || isPagePlaylist()) {
         }),
     ]
     videoGroupList.push(new Group('video-right', '右侧 视频栏', rightItems))
-
-    // 评论区
-    const commentItems = [
-        // 禁用 新版评论区
-        new CheckboxItem({
-            itemID: 'video-page-disable-next-comment-abtest',
-            description: '禁用 新版评论区, 恢复旧版 (临时功能)',
-            enableFunc: async () => {
-                let origValue = unsafeWindow.__INITIAL_STATE__
-                if (origValue?.abtest) {
-                    origValue.abtest.comment_next_version = 'DEFAULT'
-                }
-                Object.defineProperty(unsafeWindow, '__INITIAL_STATE__', {
-                    get() {
-                        return origValue
-                    },
-                    set(value) {
-                        if (value.abtest) {
-                            value.abtest.comment_next_version = 'DEFAULT'
-                        }
-                        origValue = value
-                    },
-                })
-            },
-        }),
-        // 隐藏 投票栏 (红方/蓝方)
-        // 测试视频：https://www.bilibili.com/video/av1805762267
-        new CheckboxItem({
-            itemID: 'video-page-hide-top-vote-card',
-            description: '隐藏 投票栏 (红方/蓝方)',
-            itemCSS: `.top-vote-card {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comments-header-renderer',
-                    'video-page-hide-top-vote-card',
-                    `#vote {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comments-header-renderer', 'video-page-hide-top-vote-card')
-            },
-        }),
-        // 隐藏 活动/notice, 默认开启
-        new CheckboxItem({
-            itemID: 'video-page-hide-reply-notice',
-            description: '隐藏 活动/notice',
-            defaultStatus: true,
-            itemCSS: `.reply-header .reply-notice {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comments-header-renderer',
-                    'video-page-hide-reply-notice',
-                    `#notice {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comments-header-renderer', 'video-page-hide-reply-notice')
-            },
-        }),
-        // 隐藏 评论编辑器
-        new CheckboxItem({
-            itemID: 'video-page-hide-main-reply-box',
-            description: '隐藏 评论编辑器',
-            // 不可使用display: none, 会使底部吸附评论框宽度变化
-            itemCSS: `.main-reply-box {height: 0 !important; visibility: hidden !important;}
-                .comment-container .reply-list {margin-top: -20px !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comments-header-renderer',
-                    'video-page-hide-main-reply-box',
-                    `#commentbox bili-comment-box {display: none !important;}
-                     #navbar {margin-bottom: 0 !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comments-header-renderer', 'video-page-hide-main-reply-box')
-            },
-        }),
-        // 隐藏 评论编辑器内占位文字, 默认开启
-        new CheckboxItem({
-            itemID: 'video-page-hide-reply-box-textarea-placeholder',
-            description: '隐藏 评论编辑器内占位文字',
-            defaultStatus: true,
-            itemCSS: `.main-reply-box .reply-box-textarea::placeholder {color: transparent !important; user-select: none;}
-                .fixed-reply-box .reply-box-textarea::placeholder {color: transparent !important; user-select: none;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-textarea',
-                    'video-page-hide-reply-box-textarea-placeholder',
-                    `textarea:not([placeholder^="回复"])::placeholder {color: transparent !important; user-select: none;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-textarea', 'video-page-hide-reply-box-textarea-placeholder')
-            },
-        }),
-        // 隐藏 页面底部 吸附评论框, 默认开启
-        new CheckboxItem({
-            itemID: 'video-page-hide-fixed-reply-box',
-            description: '隐藏 页面底部 吸附评论框',
-            defaultStatus: true,
-            itemCSS: `.fixed-reply-box {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comments-header-renderer',
-                    'video-page-hide-fixed-reply-box',
-                    `.bili-comments-bottom-fixed-wrapper {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comments-header-renderer', 'video-page-hide-fixed-reply-box')
-            },
-        }),
-        // 隐藏 用户卡片
-        new CheckboxItem({
-            itemID: 'video-page-hide-comment-user-card',
-            description: '隐藏 用户卡片\n鼠标放在用户名上时不显示卡片',
-            itemCSS: `.user-card {display: none!important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-user-profile',
-                    'video-page-hide-comment-user-card',
-                    `#wrap {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-user-profile', 'video-page-hide-comment-user-card')
-            },
-        }),
-        // 隐藏 评论右侧装饰
-        new CheckboxItem({
-            itemID: 'video-page-hide-reply-decorate',
-            description: '隐藏 评论右侧装饰',
-            itemCSS: `.reply-decorate {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-renderer',
-                    'video-page-hide-reply-decorate',
-                    `#ornament {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-renderer', 'video-page-hide-reply-decorate')
-            },
-        }),
-        // 隐藏 粉丝牌
-        new CheckboxItem({
-            itemID: 'video-page-hide-fan-badge',
-            description: '隐藏 粉丝牌',
-            itemCSS: `.fan-badge {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-user-medal',
-                    'video-page-hide-fan-badge',
-                    `#fans {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-user-medal', 'video-page-hide-fan-badge')
-            },
-        }),
-        // 隐藏 老粉、原始粉丝Tag
-        // 测试视频：https://www.bilibili.com/video/av479061422
-        new CheckboxItem({
-            itemID: 'video-page-hide-contractor-box',
-            description: '隐藏 老粉、原始粉丝Tag',
-            itemCSS: `.contractor-box {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-user-medal',
-                    'video-page-hide-contractor-box',
-                    `#contractor {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-user-medal', 'video-page-hide-contractor-box')
-            },
-        }),
-        // 隐藏 一级评论用户等级
-        new CheckboxItem({
-            itemID: 'video-page-hide-user-level',
-            description: '隐藏 一级评论用户等级',
-            itemCSS: `.user-level {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-user-info',
-                    'video-page-hide-user-level',
-                    `#user-level {display: none !important;}
-                    #user-name {margin-right: 5px;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-user-info', 'video-page-hide-user-level')
-            },
-        }),
-        // 隐藏 二级评论用户等级
-        new CheckboxItem({
-            itemID: 'video-page-hide-sub-user-level',
-            description: '隐藏 二级评论用户等级',
-            itemCSS: `.sub-user-level {display: none !important;}`,
-        }),
-        // 隐藏 用户头像饰品
-        new CheckboxItem({
-            itemID: 'video-page-hide-bili-avatar-pendent-dom',
-            description: '隐藏 用户头像饰品',
-            itemCSS: `.root-reply-avatar .bili-avatar-pendent-dom {display: none !important;}
-            .comment-container .root-reply-avatar .bili-avatar {width: 48px !important; height:48px !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-avatar',
-                    'video-page-hide-bili-avatar-pendent-dom',
-                    `.layer.center[style*="66px"] {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-avatar', 'video-page-hide-bili-avatar-pendent-dom')
-            },
-        }),
-        // 隐藏 用户头像徽章
-        new CheckboxItem({
-            itemID: 'video-page-hide-bili-avatar-nft-icon',
-            description: '隐藏 用户头像徽章',
-            itemCSS: `.bili-avatar-nft-icon {display: none !important;}
-                .comment-container .bili-avatar-icon {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-avatar',
-                    'video-page-hide-bili-avatar-nft-icon',
-                    `.layer:not(.center) {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-avatar', 'video-page-hide-bili-avatar-nft-icon')
-            },
-        }),
-        // 隐藏 用户投票 (红方/蓝方)
-        // 测试视频：https://www.bilibili.com/video/av1805762267
-        new CheckboxItem({
-            itemID: 'video-page-hide-vote-info',
-            description: '隐藏 用户投票 (红方/蓝方)',
-            itemCSS: `.vote-info {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-renderer',
-                    'video-page-hide-vote-info',
-                    `bili-comment-vote-option {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-renderer', 'video-page-hide-vote-info')
-            },
-        }),
-        // 隐藏 评论内容下Tag(UP觉得很赞)
-        new CheckboxItem({
-            itemID: 'video-page-hide-reply-tag-list',
-            description: '隐藏 评论内容下Tag(UP觉得很赞)',
-            itemCSS: `.reply-tag-list {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-renderer',
-                    'video-page-hide-reply-tag-list',
-                    `#tags {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-renderer', 'video-page-hide-reply-tag-list')
-            },
-        }),
-        // 隐藏 笔记评论前的小Logo, 默认开启
-        new CheckboxItem({
-            itemID: 'video-page-hide-note-prefix',
-            description: '隐藏 笔记评论前的小Logo',
-            defaultStatus: true,
-            itemCSS: `.note-prefix {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-renderer',
-                    'video-page-hide-note-prefix',
-                    `#note {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-renderer', 'video-page-hide-note-prefix')
-            },
-        }),
-        // 禁用 评论内容搜索关键词高亮, 默认开启
-        // 测试视频：https://www.bilibili.com/video/av1406084552
-        new CheckboxItem({
-            itemID: 'video-page-hide-jump-link-search-word',
-            description: '禁用 评论内容搜索关键词高亮',
-            defaultStatus: true,
-            itemCSS: `.reply-content .jump-link.search-word {color: inherit !important;}
-                .comment-container .reply-content .jump-link.search-word:hover {color: #008AC5 !important;}
-                .comment-container .reply-content .icon.search-word {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-rich-text',
-                    'video-page-hide-jump-link-search-word',
-                    `#contents a[href*="search.bilibili.com"] {color: inherit !important;}
-                    #contents a[href*="search.bilibili.com"]:hover {color: #008AC5 !important;}
-                    #contents a[href*="search.bilibili.com"] img {display: none !important;}
-                    `,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-rich-text', 'video-page-hide-jump-link-search-word')
-            },
-        }),
-        // 禁用 评论中的@高亮
-        new CheckboxItem({
-            itemID: 'video-page-hide-reply-content-user-highlight',
-            description: '禁用 评论中的@高亮',
-            itemCSS: `.sub-reply-container .reply-content .jump-link.user {color: inherit !important;}
-                .comment-container .sub-reply-container .reply-content .jump-link.user:hover {color: #40C5F1 !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-rich-text',
-                    'video-page-hide-reply-content-user-highlight',
-                    `#contents a[href*="space.bilibili.com"] {color: inherit !important;}
-                    #contents a[href*="space.bilibili.com"]:hover {color: #008AC5 !important;}
-                    `,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-rich-text', 'video-page-hide-reply-content-user-highlight')
-            },
-        }),
-        // 隐藏 召唤AI机器人的评论, 默认开启
-        new CheckboxItem({
-            itemID: 'video-page-hide-at-reply-at-bots',
-            description: '隐藏 召唤AI机器人的评论 (旧版)',
-            defaultStatus: true,
-            itemCSS:
-                // 8455326 @机器工具人
-                // 234978716 @有趣的程序员
-                // 1141159409 @AI视频小助理
-                // 437175450 @AI视频小助理总结一下 (误伤)
-                // 1692825065 @AI笔记侠
-                // 690155730 @AI视频助手
-                // 689670224 @哔哩哔理点赞姬
-                // 3494380876859618 @课代表猫
-                // 1168527940 @AI课代表呀
-                // 439438614 @木几萌Moe
-                // 1358327273 @星崽丨StarZai
-                // 3546376048741135 @AI沈阳美食家
-                // 1835753760 @AI识片酱
-                // 9868463 @AI头脑风暴
-                // 358243654 @GPT_5
-                // 393788832 @Juice_AI
-                // 91394217 @AI全文总结
-                // 473018527 @AI视频总结
-                // 3546639035795567 @AI总结视频
-                `.reply-item:has(.jump-link.user:is(
-                    [data-user-id="8455326"],
-                    [data-user-id="234978716"],
-                    [data-user-id="1141159409"],
-                    [data-user-id="437175450"],
-                    [data-user-id="1692825065"],
-                    [data-user-id="690155730"],
-                    [data-user-id="689670224"],
-                    [data-user-id="3494380876859618"],
-                    [data-user-id="1168527940"],
-                    [data-user-id="439438614"],
-                    [data-user-id="1358327273"],
-                    [data-user-id="3546376048741135"],
-                    [data-user-id="1835753760"],
-                    [data-user-id="9868463"],
-                    [data-user-id="358243654"],
-                    [data-user-id="393788832"],
-                    [data-user-id="91394217"],
-                    [data-user-id="473018527"],
-                    [data-user-id="3546639035795567"]
-                )) {
-                    display: none !important;
-                }`,
-        }),
-        // 隐藏 AI机器人发布的评论
-        new CheckboxItem({
-            itemID: 'video-page-hide-bots-reply',
-            description: '隐藏 AI机器人发布的评论 (旧版)',
-            defaultStatus: false,
-            itemCSS:
-                // 8455326 @机器工具人
-                // 234978716 @有趣的程序员
-                // 1141159409 @AI视频小助理
-                // 437175450 @AI视频小助理总结一下 (误伤)
-                // 1692825065 @AI笔记侠
-                // 690155730 @AI视频助手
-                // 689670224 @哔哩哔理点赞姬
-                // 3494380876859618 @课代表猫
-                // 1168527940 @AI课代表呀
-                // 439438614 @木几萌Moe
-                // 1358327273 @星崽丨StarZai
-                // 3546376048741135 @AI沈阳美食家
-                // 9868463 @AI头脑风暴
-                // 358243654 @GPT_5
-                // 393788832 @Juice_AI
-                // 91394217 @AI全文总结
-                // 473018527 @AI视频总结
-                // 3546639035795567 @AI总结视频
-                // 普通视频页不隐藏 1835753760 @AI识片酱 发布的评论
-                `.reply-item:has(.root-reply-container .user-name:is(
-                    [data-user-id="8455326"],
-                    [data-user-id="234978716"],
-                    [data-user-id="1141159409"],
-                    [data-user-id="437175450"],
-                    [data-user-id="1692825065"],
-                    [data-user-id="690155730"],
-                    [data-user-id="689670224"],
-                    [data-user-id="3494380876859618"],
-                    [data-user-id="1168527940"],
-                    [data-user-id="439438614"],
-                    [data-user-id="1358327273"],
-                    [data-user-id="3546376048741135"],
-                    [data-user-id="9868463"],
-                    [data-user-id="358243654"],
-                    [data-user-id="393788832"],
-                    [data-user-id="91394217"],
-                    [data-user-id="473018527"],
-                    [data-user-id="3546639035795567"]
-                )) {
-                    display: none !important;
-                }`,
-        }),
-        // 隐藏 包含@的 无人点赞评论
-        new CheckboxItem({
-            itemID: 'video-page-hide-zero-like-at-reply',
-            description: '隐藏 包含@的 无人点赞评论 (旧版)',
-            itemCSS: `.reply-item:has(.root-reply .jump-link.user):not(:has(.delete-reply, .top-icon, .sub-up-icon, .reply-info .reply-like span)) {display: none !important;}`,
-        }),
-        // 隐藏 包含@的 全部评论
-        new CheckboxItem({
-            itemID: 'video-page-hide-at-reply-all',
-            description: '隐藏 包含@的 全部评论 (旧版)',
-            itemCSS: `.reply-item:has(.root-reply .jump-link.user):not(:has(.delete-reply, .top-icon, .sub-up-icon)) {display: none !important;}`,
-        }),
-        // 隐藏 LV1 无人点赞评论
-        new CheckboxItem({
-            itemID: 'video-page-hide-zero-like-lv1-reply',
-            description: '隐藏 LV1 无人点赞评论 (旧版)',
-            itemCSS: `.reply-item:has(.user-level.level-1):not(:has(.delete-reply, .top-icon, .sub-up-icon, .reply-info .reply-like span)) {display: none !important;}`,
-        }),
-        // 隐藏 LV2 无人点赞评论
-        new CheckboxItem({
-            itemID: 'video-page-hide-zero-like-lv2-reply',
-            description: '隐藏 LV2 无人点赞评论 (旧版)',
-            itemCSS: `.reply-item:has(.user-level.level-2):not(:has(.delete-reply, .top-icon, .sub-up-icon, .reply-info .reply-like span)) {display: none !important;}`,
-        }),
-        // 隐藏 LV3 无人点赞评论
-        new CheckboxItem({
-            itemID: 'video-page-hide-zero-like-lv3-reply',
-            description: '隐藏 LV3 无人点赞评论 (旧版)',
-            itemCSS: `.reply-item:has(.user-level.level-3):not(:has(.delete-reply, .top-icon, .sub-up-icon, .reply-info .reply-like span)) {display: none !important;}`,
-        }),
-        // 一级评论 踩/回复 只在hover时显示, 默认开启
-        new CheckboxItem({
-            itemID: 'video-page-hide-root-reply-dislike-reply-btn',
-            description: '一级评论 踩/回复 只在hover时显示',
-            defaultStatus: true,
-            itemCSS: `
-                .reply-item:not(:has(i.disliked)) :is(.reply-btn, .reply-dislike) {
-                    opacity: 0;
-                }
-                @keyframes appear {
-                    0% {opacity: 0;}
-                    100% {opacity: 1;}
-                }
-                .reply-item:hover :is(.reply-btn, .reply-dislike) {
-                    animation: appear;
-                    animation-duration: 0.2s;
-                    animation-delay: 0.3s;
-                    animation-fill-mode: forwards;
-                }`,
-            enableFunc: () => {
-                /* 借用more button的display样式，改为传透明度值 */
-                shadow.register(
-                    'bili-comment-renderer',
-                    'video-page-hide-root-reply-dislike-reply-btn',
-                    `#body {
-                        --bili-comment-hover-more-display: 0 !important;
-                    }
-                    #body:hover {
-                        --bili-comment-hover-more-display: 1 !important;
-                    }`,
-                )
-                shadow.register(
-                    'bili-comment-action-buttons-renderer',
-                    'video-page-hide-root-reply-dislike-reply-btn',
-                    `#dislike button, #reply button, #more button {
-                        display: block !important;
-                        opacity: var(--bili-comment-action-buttons-more-display);
-                        transition: opacity 0.2s 0.3s;
-                    }`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-renderer', 'video-page-hide-root-reply-dislike-reply-btn')
-                shadow.unregister(
-                    'bili-comment-action-buttons-renderer',
-                    'video-page-hide-root-reply-dislike-reply-btn',
-                )
-            },
-        }),
-        // 二级评论 踩/回复 只在hover时显示, 默认开启
-        new CheckboxItem({
-            itemID: 'video-page-hide-sub-reply-dislike-reply-btn',
-            description: '二级评论 踩/回复 只在hover时显示',
-            defaultStatus: true,
-            itemCSS: `
-                .sub-reply-item:not(:has(i.disliked)) :is(.sub-reply-btn, .sub-reply-dislike) {
-                    opacity: 0;
-                }
-                @keyframes appear {
-                    0% {opacity: 0;}
-                    100% {opacity: 1;}
-                }
-                .sub-reply-item:hover :is(.sub-reply-btn, .sub-reply-dislike) {
-                    animation: appear;
-                    animation-duration: 0.2s;
-                    animation-delay: 0.3s;
-                    animation-fill-mode: forwards;
-                }`,
-            enableFunc: () => {
-                /* 借用more button的display样式，改为传透明度值 */
-                shadow.register(
-                    'bili-comment-reply-renderer',
-                    'video-page-hide-sub-reply-dislike-reply-btn',
-                    `
-                    #body {
-                        --bili-comment-hover-more-display: 0 !important;
-                    }
-                    #body:hover {
-                        --bili-comment-hover-more-display: 1 !important;
-                    }`,
-                )
-                shadow.register(
-                    'bili-comment-action-buttons-renderer',
-                    'video-page-hide-sub-reply-dislike-reply-btn',
-                    `#dislike button, #reply button, #more button {
-                        display: block !important;
-                        opacity: var(--bili-comment-action-buttons-more-display);
-                        transition: opacity 0.2s 0.3s;
-                    }`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-reply-renderer', 'video-page-hide-sub-reply-dislike-reply-btn')
-                shadow.unregister('bili-comment-action-buttons-renderer', 'video-page-hide-sub-reply-dislike-reply-btn')
-            },
-        }),
-        // 隐藏 大表情
-        // 测试视频：https://www.bilibili.com/video/av1406211916
-        new CheckboxItem({
-            itemID: 'video-page-hide-emoji-large',
-            description: '隐藏 大表情',
-            itemCSS: `.emoji-large {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-rich-text',
-                    'video-page-hide-emoji-large',
-                    `#contents img:is(.emoji-large, [style^="width:50px"]) {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-rich-text', 'video-page-hide-emoji-large')
-            },
-        }),
-        // 大表情变成小表情
-        // 测试视频：https://www.bilibili.com/video/av1406211916
-        new CheckboxItem({
-            itemID: 'video-page-hide-emoji-large-zoom',
-            description: '大表情变成小表情',
-            itemCSS: `.emoji-large {zoom: 0.5;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-rich-text',
-                    'video-page-hide-emoji-large-zoom',
-                    `#contents img:is(.emoji-large, [style^="width:50px"]) {zoom: 0.5 !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-rich-text', 'video-page-hide-emoji-large-zoom')
-            },
-        }),
-        // 用户名 全部大会员色
-        new CheckboxItem({
-            itemID: 'video-page-reply-user-name-color-pink',
-            description: '用户名 全部大会员色',
-            itemCSS: `.reply-item .user-name, .comment-container .reply-item .sub-user-name {color: #FB7299 !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-user-info',
-                    'video-page-reply-user-name-color-pink',
-                    `#user-name {color: #FB7299 !important;}
-                    #user-name a {color: #FB7299 !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-user-info', 'video-page-reply-user-name-color-pink')
-            },
-        }),
-        // 用户名 全部恢复默认色
-        new CheckboxItem({
-            itemID: 'video-page-reply-user-name-color-default',
-            description: '用户名 全部恢复默认色',
-            itemCSS: `.reply-item .user-name, .comment-container .reply-item .sub-user-name {color: #61666d !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-comment-user-info',
-                    'video-page-reply-user-name-color-default',
-                    `#user-name {color: #61666d !important;}
-                    #user-name a {color: #61666d !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-comment-user-info', 'video-page-reply-user-name-color-default')
-            },
-        }),
-        // 笔记图片 查看大图优化, 默认开启
-        // 测试视频：https://www.bilibili.com/video/av34205337
-        new CheckboxItem({
-            itemID: 'video-page-reply-view-image-optimize',
-            description: '笔记图片 查看大图优化',
-            defaultStatus: true,
-            // 单图模式隐藏底部图片列表, 多图模式淡化列表, hover复原, 左右按钮增大
-            itemCSS: `.reply-view-image .last-image, .reply-view-image .next-image {zoom: 1.4;}
-                .reply-view-image:has(.preview-item-box:only-child) .last-image {display: none !important;}
-                .reply-view-image:has(.preview-item-box:only-child) .next-image {display: none !important;}
-                .reply-view-image .preview-list {display: none !important;}`,
-            enableFunc: () => {
-                shadow.register(
-                    'bili-photoswipe',
-                    'video-page-reply-view-image-optimize',
-                    `#wrap:has(#thumb:empty) :is(#prev, #next) {display: none !important;}
-                    #prev, #next {zoom: 1.3;}
-                    #thumb {display: none !important;}`,
-                )
-            },
-            disableFunc: () => {
-                shadow.unregister('bili-photoswipe', 'video-page-reply-view-image-optimize')
-            },
-        }),
-        // 隐藏 整个评论区 #42
-        new CheckboxItem({
-            itemID: 'video-page-hide-comment',
-            description: '隐藏 整个评论区',
-            itemCSS: `#comment, #comment-module {display: none;}`,
-        }),
-    ]
-    videoGroupList.push(new Group('video-comment', '评论区', commentItems))
 
     // 右下角
     const sidebarItems = [
